requirements.txt
streamlit
pandas
google-generativeai
PyPDF2
requests
beautifulsoup4